camera

+++ pax/transform.odin
@@ -2,6 +2,13 @@ package pax
 
 Transform :: struct
 {
-    point: [2]int,
+    //
+    //
+    //
+    point: [2]f32,
+
+    //
+    //
+    //
     scale: [2]f32,
 }
